pyproject.toml
[project]
name = "satellite"
version = "0.0.6"
authors = [
  { name="Tom Young" },
]
description = "Create, update and delete rows in a synthetic EMAP star schema"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3"
]
dependencies = [
    "Faker==15.3.*",
    "GitPython==3.1.*",
    "coloredlogs==15.0.*",
    "networkx==2.8.*",
    "click==8.1.*"
]

[project.scripts]
satellite = "satellite.main:cli"

[tool.setuptools]
packages = ["satellite"]

[project.urls]
"Homepage" = "https://github.com/UCLH-DIF/Satellite"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
